releases: document `releases/latest` branch
We got rid of `master` after #17377, but users still want a way to get
the latest stable release without knowing its number.

We've added a `releases/latest` branch to replace what was once `master`.
Note that unlike `master` in Git Flow, `releases/latest` does *not* have
tags on it. It's just a pointer to the latest release, and tags live on
release branches.